Issue #4 Ability to upload images to threads posts
It's not perfect, but it's now possible to upload images. It's using
MongoDB's GridFS system. Because Mongoose does not have support for it,
we're using the native-mongodb object  retrieved from Mongoose.
There are still some things to be done, like maybe stream the file instead of
sending it as a blob to the remote. Other things are maybe better error
handling and better and more interactive viewing of the picture in the
feed.